requirements.txt
beautifulsoup4==4.12.0
biopython==1.83
EbookLib==0.18
faster_whisper==1.0.1
ffmpeg_python==0.2.0
rapidfuzz==3.8.1
numpy==1.24.3
openai_whisper==20231117
regex==2023.12.25
tqdm>=4.66.3
wcwidth>=0.2.13
onnxruntime>=1.18.0
